gigshield_v3/gigshield_v3/requirements.txt
Flask==3.1.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
Pillow==10.3.0
requests==2.32.3
gunicorn==22.0.0
python-dotenv==1.0.1
